Fixes from Artur, i.e. TTBar sample recognition

+++ HTTAnalysis/HTTHistogramsTT.cc
@@ -443,7 +443,7 @@ void HTTHistogramsTT::finalizeHistograms(int nRuns, float weight){
   ttScale = 1.0;//MB
 
   for(unsigned int iCategory = (int)HTTAnalyzerTT::jet0;
-      iCategory<(int)HTTAnalyzerTT::DUMMY;++iCategory){
+      iCategory<(int)(int)HTTAnalyzerTT::DUMMY;++iCategory){
     
     wselOSCorrection =  std::pair<float,float>(1.0,0);
     wselSSCorrection =  std::pair<float,float>(1.0,0);
@@ -452,7 +452,7 @@ void HTTHistogramsTT::finalizeHistograms(int nRuns, float weight){
     //wselSSCorrection = getWNormalisation(iCategory, "SS");
 
     //plotStack(iCategory, "MassSV");
-    plotStack(iCategory, "MassVis");   
+    plotStack(iCategory, "MassVis");
     plotStack(iCategory, "MassTrans");
     
     plotStack(iCategory, "PtMuon");
